ai/AI_verdict/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
google-genai>=0.3.0
openai==1.12.0

python-dotenv==1.0.1

chromadb==0.4.24

httpx==0.27.0
